wait before attempting to delete k8s namespace
Wait for before trying to delete the k8s namespace on cluster down and
when MNG and the ALB/NLB testers have been run.

Because we asked Kubernetes to delete the NLB hello world and ALB2048
Deployment/Service above, and both of these interact with the underlying
Kubernetes AWS cloud provider to clean up the cloud load balancer
backing the Service of type LoadBalancer. The calls to delete the
Service return immediately (successfully) but the cloud load balancer
resources may not have been deleted yet, including the ENIs that were
associated with the cloud load balancer. When, later, aws-k8s-tester
tries deleting the VPC associated with the test cluster, it will run
into permissions issues because the IAM role that created the ENIs
associated with the ENIs in subnets associated with the cloud load
balancers will no longer exist.

Issue #70