ensure sub-tester instances not-nil
When tearing down Kubernetes resources created during the job
sub-testers (e.g. NLB hello world), Tester.down() was skipping the
deletion of the resources when `aws-k8s-tester delete cluster` was being
run. This was due to the construction of the sub-tester instances only
happening when Tester.up() was called.

This patch fixes this by always creating the sub-tester instances on
construction of the main Tester object.

Issue #70